requirements.txt
asgiref==3.4.0
crispy-bootstrap5==0.4
Django==3.2.4
django-crispy-forms==1.12.0
django-environ==0.4.5
django-ratelimit==3.0.1
djangorestframework==3.12.4
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
